// src/regions/dto/public-region-response.dto.ts
export class PublicRegionResponseDto {
  id: number;
  uuid: string;
  regionName: string;
  regionNameHi: string | null;
  regionCode: string;
  dcRoName: string | null;
  dcRoNameHi: string | null;
  address: string | null;
  addressHindi: string | null;
  phone: string | null;
  email: string | null;
}